+++ pyteal/ast/__init__.py
@@ -42,6 +42,7 @@ from pyteal.ast.array import Array
 from pyteal.ast.tmpl import Tmpl
 from pyteal.ast.nonce import Nonce
 from pyteal.ast.pragma import Pragma
+from pyteal.ast.comment import Comment
 
 # unary ops
 from pyteal.ast.unaryexpr import (
@@ -216,6 +217,7 @@ __all__ = [
     "Tmpl",
     "Nonce",
     "Pragma",
+    "Comment",
     "UnaryExpr",
     "Btoi",
     "Itob",