+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Unreleased
 
+## Added
+
+* Static and Dynamic Byte Array convenience classes ([#500](https://github.com/algorand/pyteal/pull/500))
+
+## Fixed
+
+
+## Changed
+
+
 # 0.16.0
 
 ## Added
@@ -7,7 +17,6 @@
 * Add the ability to pass foreign reference arrays directly into inner transactions ([#384](https://github.com/algorand/pyteal/pull/384))
 
 * NamedTuple Implementation ([#473](https://github.com/algorand/pyteal/pull/473))
-* Static and Dynamic Byte Array convenience classes ([#500](https://github.com/algorand/pyteal/pull/500))
 
 * ExecuteMethodCall helper ([#501](https://github.com/algorand/pyteal/pull/501))