+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Unreleased
 
 ## Fixed
-* TODO
+* Fix AST duplication bug in `String.set` when called with an `Expr` argument ([#508](https://github.com/algorand/pyteal/pull/508))
 
 # 0.16.0