+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Unreleased
 
+## Added
+* Add the ability to insert comments in TEAL source file with the Comment method ([#410](https://github.com/algorand/pyteal/pull/410))
+
+## Fixed
+## Changed
+
 # 0.16.0
 
 ## Added